requirements.txt
streamlit==1.33.0
google-generativeai==0.5.2
python-dotenv==1.0.1
pymongo==4.6.3
pandas==2.2.2
pyparsing==3.1.2
plotly==5.22.0
requests==2.32.3
